+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added --peer-slots to "rg c", "rg m" and "rg spawn"
+
 ## [1.20.1] - 2023-10-25
 
 ### Added